requirements.txt
discord.py >= 1.7.3
flask >= 2.0.1
apscheduler >= 3.7.0
Flask-APScheduler >= 1.12.2
requests
python-dotenv
flask-restful
flask-cors